Complete rewrite - KeyTable 2.0
Focusing on using the DataTables 1.10 API and full integraiton with the
new DataTables framework build structure, this update is a complete
revamp of Editor for a more "modern" style of working with DataTables
and Javascript in general.

It basically implements all of the features of v1, but it now requires
that the table be a DataTable. It will work with server-side processing
correctly now and the code size is much smaller.

There are still a few areas that need work and documentation is
required, but the basic funtionality is now there.