Fixed errors caused by focusing filtered out rows.
Changing the search filter while a cell is focused, in such a way that the cell's row gets filtered out, leaves the saved state pointing to a row that was filtered out. When such a state is loaded, the attempt to focus the row that was filtered out causes errors. This change should simply ignore such invalid states.